imported the style from my toshokan project.
I also opted to switch to bower. I am fully aware that it is considered "deprecated", but npm doesn't seem to play nice with flask and I don't want to waste too much time looking into other solutions right now. A lot of style classes and various text will of course have to be renamed, but I feel like this is a great start for the "look" of the site.