added jquery-ui to the compiled javascript.
I'm not sure if it's working, so I might have to remove it. I also removed vue-devtools because it didn't like being compiled with the other libraries, which is a necessity since you can't move the node_modules folder. If I want to use vue-devtools in the future, I should just use a CDN link.