the login form now doesn't act crazy if you click too much.
for some reason, underscore's throttle function wasn't working for me. I ended up implementing a lock that basically prevents the toggle from happening again until the animation is over. It does make the login form a bit more rigid, but I think that's a fair tradeoff for not opening and closing repeatiately if you spam click it.