began implementing remember me functionality.
the site already had this, but it assumed you wanted to be remembered. i'm working on adding the choice. i removed vue-ls because although localStorage is cool, it's for Javascript only. so in my case it makes more sense to use a cookie because both Javascript and Python can interact with that. I currently have the site set to keep you logged in for 7 days if you check the box. the vue-cookie plugin isn't working currently, so this commit is broken.