removed migrations.
this is kind of a large, sweeping change to make in this branch. But this migration system isn't working out and it would only get worse going forward. any time I need to make a "migration", I'll probably have to write a sql script myself or find a better migration library.